requirements.txt
Django==4.2.23
django-debug-toolbar
psycopg2-binary
dotenv
python-dotenv
django-environ